Create Vehicle Alarm security
It's an IoT automotive secure alarm system. When a vehicle is parked, the user will get to see the pictures captured by four cameras from each corner of the car when there are potential threats. Ideally, this information will sent to users phone using ReST server. This is the program, which will be processed before the server step.